+++ cf/data/fragment/__init__.py
@@ -1,5 +1,2 @@
-from .fullfragmentarray import FullFragmentArray
-from .h5netcdffragmentarray import H5netcdfFragmentArray
-from .netcdffragmentarray import NetCDFFragmentArray
-from .netcdf4fragmentarray import NetCDF4FragmentArray
-from .umfragmentarray import UMFragmentArray
+from .fragmentfilearray import FragmentFileArray
+from .fragmentumarray import FragmentUMArray